modules/fmops/qna-rag/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
attrs==23.2.0
    # via
    #   -r requirements.in
    #   cattrs
    #   jsii
aws-cdk-asset-awscli-v1==2.2.202
    # via aws-cdk-lib
aws-cdk-asset-kubectl-v20==2.1.2
    # via aws-cdk-lib
aws-cdk-asset-node-proxy-agent-v6==2.0.3
    # via aws-cdk-lib
aws-cdk-lib==2.137.0
    # via
    #   -r requirements.in
    #   cdk-nag
    #   cdklabs-generative-ai-cdk-constructs
boto3==1.34.108
    # via -r requirements.in
botocore==1.34.108
    # via
    #   boto3
    #   s3transfer
cattrs==23.2.3
    # via jsii
cdk-nag==2.28.118
    # via cdklabs-generative-ai-cdk-constructs
cdklabs-generative-ai-cdk-constructs==0.1.145
    # via -r requirements.in
constructs==10.3.0
    # via
    #   aws-cdk-lib
    #   cdk-nag
    #   cdklabs-generative-ai-cdk-constructs
importlib-resources==6.4.0
    # via jsii
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
jsii==1.98.0
    # via
    #   aws-cdk-asset-awscli-v1
    #   aws-cdk-asset-kubectl-v20
    #   aws-cdk-asset-node-proxy-agent-v6
    #   aws-cdk-lib
    #   cdk-nag
    #   cdklabs-generative-ai-cdk-constructs
    #   constructs
publication==0.0.3
    # via
    #   aws-cdk-asset-awscli-v1
    #   aws-cdk-asset-kubectl-v20
    #   aws-cdk-asset-node-proxy-agent-v6
    #   aws-cdk-lib
    #   cdk-nag
    #   cdklabs-generative-ai-cdk-constructs
    #   constructs
    #   jsii
python-dateutil==2.9.0.post0
    # via
    #   botocore
    #   jsii
s3transfer==0.10.1
    # via boto3
six==1.16.0
    # via python-dateutil
typeguard==2.13.3
    # via
    #   aws-cdk-asset-awscli-v1
    #   aws-cdk-asset-kubectl-v20
    #   aws-cdk-asset-node-proxy-agent-v6
    #   aws-cdk-lib
    #   cdk-nag
    #   cdklabs-generative-ai-cdk-constructs
    #   constructs
    #   jsii
typing-extensions==4.11.0
    # via jsii
urllib3==1.26.17
    # via botocore
